src/requirements.txt
google-auth
gspread
Flask==3.0.0
gunicorn==20.1.0
Werkzeug==3.0.1
gspread_dataframe
BeautifulSoup4
lxml
html5lib
pyairtable
hubspot-api-client
openai
python-docx
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
notion-client